Added single invocation geometry shader.
I haven't gotten a chance to test this on all hardware, but so far it
works with the Nvidia shader compiler (which means next to nothing,
really).

Basically, barrel.geom generates two primitives and passes along all the
proper settings with them. This allows you to apply the distortion
effect from a single shader binding/draw call and without having to
rebind it for the other eye and draw again.